rename: 파일명 변경 view/index.js -> view/elements.js
